Scrape skincare product details for the given categories:['face-moisturisers', 'cleanser', 'sunscreen', 'concealer', 'mask-and-peel', 'foundation']
Follow a two step plan. First, it obtains the URLs of the individual
products under each category. Then, it opens up the product detail page
for each of the products and scrapes off their details.

The scraper is an  efficient one and is generally able to spot the
relevant elements of the web pages without errors.